pass lint and skip UTs
